Moving F3DOptions to Application
